chore: create base file for model and update its references
